feat: Completely simplify AdminDashboard - remove project mgmt features
Simplified dashboard from complex project management tool to clean content
overview matching the purpose: strategic plan REPORTING and PRESENTATION.

Removed (Confusing/Not Needed):
- Goals Needing Attention (this isn't a project tracking tool)
- Updates Due metrics (administrators track this elsewhere)
- Recalculate Progress button (overkill, unreliable)
- System Health gradient card (project mgmt feature)
- All complex progress tracking UI

New Simplified Design:
- 3 stat cards: Strategic Objectives, Goals, Metrics (simple counts)
- 6 quick action cards: Create, Manage Goals, Manage Schools, Import, Branding, View Public
- Getting Started guide with clear steps
- ALL white cards with light icon backgrounds (no gradients)
- Clean, professional Webflow-style aesthetic

Quick Actions Now Include:
- Manage Schools (direct link to /:slug/admin/schools)
- All actions relevant to strategic plan content management
- No confusing project management features

From 438 lines → 173 lines (60% reduction)
From complex dashboard → clean, focused interface

This is a REPORTING TOOL for strategic plans, not a project management tool.
